# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Customer Service Management API",
        "description": "A comprehensive API for managing customers and their services",
        "contact": {
            "email": "admin@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/services": {
            "get": {
                "tags": [
                    "Services"
                ],
                "summary": "Get all services",
                "description": "Retrieve a paginated list of all services",
                "operationId": "7427356885cfb02aa536e9f117be4cac",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page number",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Items per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search term",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filter by status",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "active",
                                "inactive",
                                "pending",
                                "completed"
                            ]
                        }
                    },
                    {
                        "name": "type",
                        "in": "query",
                        "description": "Filter by type",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "consultation",
                                "development",
                                "maintenance",
                                "support",
                                "training",
                                "other"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Services retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Services retrieved successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "current_page": {
                                                    "type": "integer"
                                                },
                                                "total": {
                                                    "type": "integer"
                                                },
                                                "per_page": {
                                                    "type": "integer"
                                                },
                                                "data": {
                                                    "type": "array",
                                                    "items": {
                                                        "$ref": "#/components/schemas/Service"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Services"
                ],
                "summary": "Create a new service",
                "description": "Create a new service for a customer",
                "operationId": "5036c80b50b3692feb9eb282be8923bc",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "customer_id",
                                    "name",
                                    "type",
                                    "price"
                                ],
                                "properties": {
                                    "customer_id": {
                                        "description": "Customer ID",
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "name": {
                                        "description": "Service name",
                                        "type": "string",
                                        "example": "Web Development"
                                    },
                                    "description": {
                                        "description": "Service description",
                                        "type": "string",
                                        "example": "Complete website development"
                                    },
                                    "type": {
                                        "description": "Service type",
                                        "type": "string",
                                        "enum": [
                                            "consultation",
                                            "development",
                                            "maintenance",
                                            "support",
                                            "training",
                                            "other"
                                        ],
                                        "example": "development"
                                    },
                                    "price": {
                                        "description": "Service price",
                                        "type": "number",
                                        "format": "float",
                                        "example": 2500
                                    },
                                    "status": {
                                        "description": "Service status",
                                        "type": "string",
                                        "enum": [
                                            "active",
                                            "inactive",
                                            "pending",
                                            "completed"
                                        ],
                                        "example": "active"
                                    },
                                    "start_date": {
                                        "description": "Start date",
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024-02-01"
                                    },
                                    "end_date": {
                                        "description": "End date",
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024-03-15"
                                    },
                                    "duration_hours": {
                                        "description": "Duration in hours",
                                        "type": "integer",
                                        "example": 120
                                    },
                                    "features": {
                                        "description": "Service features",
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        },
                                        "example": [
                                            "Responsive Design",
                                            "SEO"
                                        ]
                                    },
                                    "notes": {
                                        "description": "Additional notes",
                                        "type": "string",
                                        "example": "Special requirements"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Service created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Service created successfully"
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Service"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Service": {
                "title": "Service",
                "description": "Service model",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "customer_id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Web Development"
                    },
                    "description": {
                        "type": "string",
                        "example": "Complete website development with modern technologies"
                    },
                    "type": {
                        "type": "string",
                        "enum": [
                            "consultation",
                            "development",
                            "maintenance",
                            "support",
                            "training",
                            "other"
                        ],
                        "example": "development"
                    },
                    "price": {
                        "type": "number",
                        "format": "float",
                        "example": 2500
                    },
                    "status": {
                        "type": "string",
                        "enum": [
                            "active",
                            "inactive",
                            "pending",
                            "completed"
                        ],
                        "example": "active"
                    },
                    "start_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2024-02-01"
                    },
                    "end_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2024-03-15"
                    },
                    "duration_hours": {
                        "type": "integer",
                        "example": 120
                    },
                    "features": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "example": [
                            "Responsive Design",
                            "SEO Optimization",
                            "Admin Panel"
                        ]
                    },
                    "notes": {
                        "type": "string",
                        "example": "Client requires mobile-first approach"
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-01T00:00:00.000000Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2024-01-01T00:00:00.000000Z"
                    },
                    "customer": {
                        "properties": {
                            "id": {
                                "type": "integer",
                                "example": 1
                            },
                            "first_name": {
                                "type": "string",
                                "example": "John"
                            },
                            "last_name": {
                                "type": "string",
                                "example": "Doe"
                            },
                            "email": {
                                "type": "string",
                                "example": "john.doe@example.com"
                            },
                            "company": {
                                "type": "string",
                                "example": "Acme Corp"
                            }
                        },
                        "type": "object"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Laravel Sanctum token authentication",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
    "security": [
        {
            "sanctum": []
        }
    ]
}